WIP: make FillArrays.jl a weakdep
